Setup Home Page and basic display of items.
Home page is visited right away, and has instructions and a cool tank gif. Also altered the database to have better Inheritance, which simplifies the number of tables down to the top-level categories. Sub categories are differentiated with a new discriminator column. Lastly, I setup basic data viewing. I still need to figure out how to use a string to specify which table to select. I.E. if I use the string "Pistols", have the code set the item source to db.Pistols, and if it is "Mortars", have it set the source to db.Mortars, and so on. Reflection might be the answer, but need more research.