Switched from Listview to Gridview for appearances
Also finally figured out how to access the clicked item in a list/grid view using the eventargs e.clickedItem. Must write switch statement to handle each category type, as the e.clickedItem object must be converted to the proper type. Lastly, figure out how I want the gridview to layout the data. Thinking of having the top row span the image, with the rest of the grid having the name of the property on the left and the corresponding value on the right, all underneath the image. There would then be two gridviews next to each other, each corresponding to the items picked for comparison.