Cleaned up code and added some Unit Tests
Got rid of unneeded commented code in various files. Also reduced the number of code-behind pages interacting directly with the database, instead, going the the DatabaseHelper class. Did some light re-organizing of files in the directory structure. Lastly, added a Unit test project and added UTs for the creation of the database, and the addition and deletion to and from the database.