Completed Comparison Operator Overloading and Display
All table items can be compared with each other, as long as they are in the same table. The User can initiate it on the compare page with a button, and a winner is displayed based on which item has more properties which are better. I.E. object 1 has 5 props better than object 2 and object 2 has 3 props better than object 1, therefore object 1 wins.