Add and Remove from database Completed.
There are now two more pages, an Insert and Delete page. The Insert Page has table and type navigation which displays a form to fill out depending on the table selected. Once submit is hit,  the item has been added to the database and can be seen on the data/compare pages. The delete page is similar to the Data page, but instead of keeping track of 2 items and passing them on to the Compare page, it instead keeps track of the selected item and when Delete is pressed, deletes the item from the database, and refreshing the displayed list.